Fix package restore namespaces
At some point a bunch of code was moved from the VS layer into the host-agnostic layer. However the namespaces were not updated, and the unit tests were not moved.

This commit finalizes that move and makes everything consistent. All these types are internal, so there are no outside dependencies for which moving these types between namespaces would be an issue.
