+++ src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/Logging/IManagedProjectDiagnosticOutputService.cs
@@ -1,6 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
-namespace Microsoft.VisualStudio.ProjectSystem.VS;
+namespace Microsoft.VisualStudio.ProjectSystem;
 
 /// <summary>
 ///   A service to print out diagnostic messages about project system functionality.
