+++ src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/PackageRestore/INuGetRestoreService.cs
@@ -1,8 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
-using Microsoft.VisualStudio.ProjectSystem.PackageRestore;
-
-namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore;
+namespace Microsoft.VisualStudio.ProjectSystem.PackageRestore;
 
 /// <summary>
 ///     Responsible for handling NuGet package restore requests.
