+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Mocks/IPackageRestoreUnconfiguredInputDataSourceFactory.cs
@@ -1,9 +1,8 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using System.Threading.Tasks.Dataflow;
-using Microsoft.VisualStudio.ProjectSystem.PackageRestore;
 
-namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore
+namespace Microsoft.VisualStudio.ProjectSystem.PackageRestore
 {
     internal static class IPackageRestoreUnconfiguredInputDataSourceFactory
     {
