+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/ProjectSystem/PackageRestore/PackageRestoreDataSourceMocked.cs
@@ -1,19 +1,18 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using Microsoft.VisualStudio.IO;
-using Microsoft.VisualStudio.ProjectSystem.PackageRestore;
 
-namespace Microsoft.VisualStudio.ProjectSystem.VS.PackageRestore.Snapshots
+namespace Microsoft.VisualStudio.ProjectSystem.PackageRestore
 {
     internal class PackageRestoreDataSourceMocked : PackageRestoreDataSource
     {
         public PackageRestoreDataSourceMocked(
             UnconfiguredProject project,
             PackageRestoreSharedJoinableTaskCollection sharedJoinableTaskCollection,
-            IPackageRestoreUnconfiguredInputDataSource dataSource, 
-            IProjectAsynchronousTasksService projectAsynchronousTasksService, 
-            IFileSystem fileSystem, 
-            IManagedProjectDiagnosticOutputService logger, 
+            IPackageRestoreUnconfiguredInputDataSource dataSource,
+            IProjectAsynchronousTasksService projectAsynchronousTasksService,
+            IFileSystem fileSystem,
+            IManagedProjectDiagnosticOutputService logger,
             INuGetRestoreService nuGetRestoreService,
             IPackageRestoreCycleDetector cycleDetector)
             : base(project, sharedJoinableTaskCollection, dataSource, projectAsynchronousTasksService, fileSystem, logger, nuGetRestoreService, cycleDetector)
