+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests.csproj
@@ -15,4 +15,8 @@
     <ProjectReference Include="..\..\src\Microsoft.VisualStudio.ProjectSystem.Managed\Microsoft.VisualStudio.ProjectSystem.Managed.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="ProjectSystem\VS\PackageRestore\Snapshots\" />
+  </ItemGroup>
+
 </Project>
